media-source/MediaSource-removeSourceBuffer-step1.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

async_test(function(t) {
    var mediaSource = new MediaSource();
    t.step(function() {
        assert_throws({name : 'InvalidAccessError'},
                      function() {
                          mediaSource.removeSourceBuffer(null);
                      },
                      'Set invalid MIME');
    });
    t.done();
}, 'MediaSource#removeSouceBuffer : 1. If sourceBuffer is null then throw an INVALID_ACCESS_ERR exception and abort these steps.');

</script>
</body>
</html>
